elibrary/requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.1.2
django-nested-admin==3.3.3
gunicorn==20.0.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
loguru==0.5.3
mysqlclient==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
xlrd==1.2.0
